% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code_for_ClinicalSig.R
\name{clinSig}
\alias{clinSig}
\title{Function for categorizing clinical significance}
\usage{
clinSig(data, pre, post, consistency, mPath, sdPath, sigLevel)
}
\arguments{
\item{data}{Data set}

\item{pre}{The variable containing the values from before the intervention}

\item{post}{The variable containing the values from after the intervention}

\item{consistency}{Estimate for the internal consistency of the used measure}

\item{mPath}{Mean score in clinical norm group}

\item{sdPath}{Standard deviation in clinical norm group}

\item{sigLevel}{The accepted probability for a type I error.}
}
\value{
Each person is assigned to one of 7 classes
}
\description{
Function for categorizing clinical significance
}

