% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code_for_ClinicalSig.R
\name{plotCs}
\alias{plotCs}
\title{Function for plotting rci and rci class}
\usage{
plotCs(data, pre, post, sigLevel, rtt, consistency, mPath, sdPath, scaleName,
  min, max)
}
\arguments{
\item{data}{Data set}

\item{pre}{The variable containing the values from before the intervention}

\item{post}{The variable containing the values from after the intervention}

\item{sigLevel}{The accepted probability for a type I error.}

\item{rtt}{Test-retest correlation for the used measure. Should be from the relevant population and with no intervention in between}

\item{consistency}{Estimate for the internal consistency of the used measure.}

\item{mPath}{Mean score in clinical norm group}

\item{sdPath}{Standard deviation in clinical norm group}

\item{scaleName}{Name of the used instrument}

\item{min}{Theoretical score minimum in instrument used}

\item{max}{Theoretical score maximum in instrument used}
}
\value{
A plot depicting the 7 classes and all CIs
}
\description{
Function for plotting rci and rci class
}

