% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code for ClinicalSig.R
\name{rci}
\alias{rci}
\title{Function for the reliable change index}
\usage{
rci(pre, post, data, rtt, sdNorm)
}
\arguments{
\item{pre}{The variable containing the values from before the intervention}

\item{post}{The variable containing the values from after the intervention}

\item{data}{Data set containing the variables}

\item{rtt}{Test-retest correlation for the used measure. Should be from the relevant population and with no intervention in between}

\item{sdNorm}{Standard deviation of the used measure in the norm group.}
}
\value{
Returns the RCI for each person in the data set
}
\description{
Function for the reliable change index
}

