% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code_for_ClinicalSig.R
\name{rciClass}
\alias{rciClass}
\title{Title}
\usage{
rciClass(data, nameRci, sigLevel)
}
\arguments{
\item{data}{Data set}

\item{nameRci}{Name of the variable in data set representing the reliable change index}

\item{sigLevel}{The accepted probability for a type I error.}
}
\value{
RCI variable is classified as changed or not
}
\description{
Title
}

