% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code for ClinicalSig.R
\name{tTest}
\alias{tTest}
\title{t test function for paired t test including Hedges g and power}
\usage{
tTest(pre, post, data, alternative, sig.level)
}
\arguments{
\item{pre}{The variable containing the values from before the intervention}

\item{post}{The variable containing the values from after the intervention}

\item{data}{The data set containing the variables}

\item{alternative}{Direction of the alternative. Can have the values "less" or "greater".}

\item{sig.level}{The accepted probability for a type I error.}
}
\value{
a list containing results from t test for dependent variables, effect size and power
}
\description{
t test function for paired t test including Hedges g and power
}
\author{
Matthias Ziegler
}

